+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/media/OptionallyStreamable.kt
@@ -1,0 +1,5 @@
+package dev.inmo.tgbotapi.types.media
+
+interface OptionallyStreamable {
+    val supportsStreaming: Boolean
+}