+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## 31.2.0
 
+* `Core`:
+  * Potentially fix [#1027](https://github.com/InsanusMokrassar/ktgbotapi/issues/1027) - drop http request exceptions on getting updates
 * `Core` + `API`:
-  * Add `supportsStreaming` in places it haven't been supported.
+  * Add `supportsStreaming` in places it haven't been supported (fix of [#1026](https://github.com/InsanusMokrassar/ktgbotapi/issues/1026))
     * **PARTIALLY BREAKING CHANGE** `supportStreaming` has been renamed to `supportsStreaming`
 
 ## 31.1.0